sprand: fix calculation of number of valid pages
Account for the case where the number of valid pages may change by more
than one from erase block to erase block.

Also, previously the remainder was kept as the integer remainder left
over after integer division. This is problematic because the denominator
is increasing in each interation. So the value of the remainder shifts
when it should not. Change this calculation to store the remainder as a
fraction of an erase block which matches the Python code.

+++ sprand.h
@@ -27,7 +27,7 @@ struct sprand_state {
 	uint64_t pagesperblock;
 	uint64_t cur_state_blocks;
 	uint64_t blocks_per_state;
-	uint64_t remainder;
+	double remainder;
 	double n_bar;
 };
